+++ python/GafferSceneUI/ShaderTweaksUI.py
@@ -127,6 +127,8 @@ Gaffer.Metadata.registerNode(
 		"tweaks.*" : [
 
 			"noduleLayout:visible", False, # Can be shown individually using PlugAdder above
+			"tweakPlugValueWidget:allowCreate", True,
+			"tweakPlugValueWidget:allowRemove", True,
 
 		],